Renaming and reorganize packages
