+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:orientation="vertical"
     tools:context="com.darwindeveloper.onecalendarview.MainActivity">
 
-    <com.darwindeveloper.onecalendar.views.OneCalendarView
+    <com.darwindeveloper.onecalendar.ui.OneCalendarView
         android:id="@+id/oneCalendar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
