+++ onecalendar/src/main/java/com/darwindeveloper/onecalendar/list/ViewHolderDay.java
@@ -1,4 +1,4 @@
-package com.darwindeveloper.onecalendar.clases;
+package com.darwindeveloper.onecalendar.list;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
